Issue resolved with gradle build where bintray and sonatype properties are not defined.
This issue would only manifest itself if the user does not have bintrayUser, bintrayKey, sonatypeUsername and sonatypepassword defined in a global gradle properties file (typically found in <user.home>. This fix tests for the properties on the gradle project file using the hasProperty method. If those properties exist, they are used, if not an empty string will be supplied to the binary plugin.

The end result is that the script now compiles, and tasks can be run (in the case where these global properties are not defined).

A reformat of the build file was also done which explains the slight formatting changes found in this change.